Relocated RPGMaster to subfolder RPGMaster/RPGMaster. Made changes to index.html and custom.css for ui changes, mainly for right character-display and child elements. Also edited schema.sql, changed back float to decimal(18,2) for consistency and better usage. Also modified StoredProcedures sql scripts to have the IF EXSITS line as mentions in r305.

+++ branches/RPGMaster/RPGSvc/SQL/Schema.sql
@@ -22,14 +22,14 @@ CREATE TABLE PlayerSkill
 PlayerSkillID INT IDENTITY(1,1),
 PlayerID INT,
 SkillID INT,
-Value decimal
+Value decimal(18,2)
 );
 CREATE TABLE PlayerStat
 (
 PlayerStatID INT IDENTITY(1,1),
 PlayerID INT,
 StatID INT,
-Value float
+Value decimal(18,2)
 );
 
 ALTER TABLE Player
